requirements.txt
ccxt
python-dotenv
requests
pyTelegramBotAPI
ta
pandas
httpx